requirements.txt
Pillow==9.5.0
pyheif==0.7.1
